feat: 메시지 기록 API 추가
- 통화 시작 시 ephemeral token을 발급받고 CallLog 초기화하는 `/call_logs/init` 추가
- 통화 중 메시지를 누적 기록하는 `/call_logs/{id}/messages` API 추가
- 기존 ConversationService 관련 클래스 제거하고 CallLogService로 통합 관리
