+++ backend/src/main/java/com/airing/backend/callLog/controller/CallLogController.java
@@ -4,6 +4,7 @@ import java.time.YearMonth;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,10 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.airing.backend.callLog.dto.CallLogDetailResponse;
-import com.airing.backend.callLog.dto.CallLogEventRequest;
+import com.airing.backend.callLog.dto.CallLogInitRequest;
+import com.airing.backend.callLog.dto.CallLogInitResponse;
 import com.airing.backend.callLog.dto.CallLogLatestResponse;
 import com.airing.backend.callLog.dto.CallLogMonthlyResponse;
 import com.airing.backend.callLog.service.CallLogService;
+import com.airing.backend.common.model.Message;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -43,6 +46,22 @@ public class CallLogController {
             @PathVariable("id") @Valid Long callLogId) {
 
         return callLogService.getCallLogDetail(userId, callLogId);
+    }
+
+    @PostMapping("/init")
+    public CallLogInitResponse initCallLog(
+            @AuthenticationPrincipal Long userId,
+            @RequestBody CallLogInitRequest request) {
+        return callLogService.initCallLog(userId, request);
+    }
+
+    @PostMapping("/{id}/messages")
+    public ResponseEntity<Void> recordCallLogMessages(
+            @AuthenticationPrincipal Long userId,
+            @PathVariable("id") @Valid Long callLogId,
+            @RequestBody List<Message> messages) {
+        callLogService.recordCallLogMessages(userId, callLogId, messages);
+        return ResponseEntity.ok().build();
     }
 
     @GetMapping
