+++ backend/src/main/java/com/airing/backend/callLog/dto/CallLogInitRequest.java
@@ -1,0 +1,14 @@
+package com.airing.backend.callLog.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class CallLogInitRequest {
+    private String callType;
+}
