+++ backend/src/main/java/com/airing/backend/callLog/dto/CallLogInitResponse.java
@@ -1,0 +1,11 @@
+package com.airing.backend.callLog.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class CallLogInitResponse {
+    private String ephemeralToken;
+    private Long callLogId;
+}
