--- backend/src/main/java/com/airing/backend/conversations/dto/ConversationInitResponse.java
@@ -1,11 +1,0 @@
-package com.airing.backend.conversations.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ConversationInitResponse {
-    private String ephemeralToken;
-    private Long conversationId;
-}
